fix: int 값을 초과하는 당첨금 수령 시 수익률이 음수로 출력되는 버그 수정

+++ src/main/java/lotto/domain/LottoGameResult.java
@@ -40,10 +40,10 @@ public class LottoGameResult {
                 .sum();
     }
 
-    private int calculateTotalPrice() {
-        int numberOfTotalLotto = resultMap.values()
+    private double calculateTotalPrice() {
+        double numberOfTotalLotto = resultMap.values()
                 .stream()
-                .mapToInt(num -> num)
+                .mapToDouble(num -> num)
                 .sum();
 
         return numberOfTotalLotto * Money.PRICE_OF_LOTTO;
